+++ src/rapidfuzz/distance/Indel.hpp
@@ -9,6 +9,11 @@
 
 namespace rapidfuzz {
 
+/**
+ * @addtogroup Distance
+ * @{
+ */
+
 template <typename InputIt1, typename InputIt2>
 size_t indel_distance(InputIt1 first1, InputIt1 last1, InputIt2 first2, InputIt2 last2,
                       size_t score_cutoff = std::numeric_limits<size_t>::max())
@@ -182,10 +187,14 @@ private:
     CachedLCSseq<CharT1> scorer;
 };
 
+#ifdef RAPIDFUZZ_DEDUCTION_GUIDES
 template <typename Sentence1>
 explicit CachedIndel(const Sentence1& s1_) -> CachedIndel<char_type<Sentence1>>;
 
 template <typename InputIt1>
 CachedIndel(InputIt1 first1, InputIt1 last1) -> CachedIndel<iter_value_t<InputIt1>>;
+#endif
+
+/** @} */
 
 } // namespace rapidfuzz