+++ src/rapidfuzz_postfix.cpp
@@ -1,6 +1,5 @@
 #include <Rcpp.h>
-#include <rapidfuzz/details/Range.hpp>
-#include <rapidfuzz/distance/Postfix_impl.hpp>
+#include <rapidfuzz/distance.hpp>
 
 #ifndef SIZE_MAX
 #define SIZE_MAX static_cast<size_t>(-1)
@@ -8,7 +7,7 @@
 
 using namespace Rcpp;
 
-//'
+//' @name postfix_distance
 //' @title Postfix Distance
 //' @description Calculates the distance between the postfixes of two strings.
 //'
@@ -20,13 +19,14 @@ using namespace Rcpp;
 //'
 //' @examples
 //' postfix_distance("string1", "string2")
+//' @export
 // [[Rcpp::export]]
 size_t postfix_distance(std::string s1, std::string s2, Rcpp::Nullable<double> score_cutoff = R_NilValue) {
   size_t cutoff_value = score_cutoff.isNull() ? std::numeric_limits<size_t>::max() : Rcpp::as<size_t>(score_cutoff);
-  size_t score_hint = std::numeric_limits<size_t>::max();
-  return rapidfuzz::detail::Postfix::distance(s1, s2, cutoff_value, score_hint);
+  return rapidfuzz::postfix_distance(s1, s2, cutoff_value);
 }
 
+//' @name postfix_similarity
 //' @title Postfix Similarity
 //' @description Calculates the similarity between the postfixes of two strings.
 //'
@@ -38,13 +38,13 @@ size_t postfix_distance(std::string s1, std::string s2, Rcpp::Nullable<double> s
 //'
 //' @examples
 //' postfix_similarity("string1", "string2")
+//' @export
 // [[Rcpp::export]]
 size_t postfix_similarity(std::string s1, std::string s2, size_t score_cutoff = 0) {
-  size_t score_hint = 0;
-  return rapidfuzz::detail::Postfix::similarity(s1, s2, score_cutoff, score_hint);
+  return rapidfuzz::postfix_similarity(s1, s2, score_cutoff);
 }
 
-
+//' @name postfix_normalized_distance
 //' @title Normalized Postfix Distance
 //' @description Calculates the normalized distance between the postfixes of two strings.
 //'
@@ -56,12 +56,13 @@ size_t postfix_similarity(std::string s1, std::string s2, size_t score_cutoff = 
 //'
 //' @examples
 //' postfix_normalized_distance("string1", "string2")
+//' @export
 // [[Rcpp::export]]
 double postfix_normalized_distance(std::string s1, std::string s2, double score_cutoff = 1.0) {
-  double score_hint = 1.0;
-  return rapidfuzz::detail::Postfix::normalized_distance(s1, s2, score_cutoff, score_hint);
+  return rapidfuzz::postfix_normalized_distance(s1, s2, score_cutoff);
 }
 
+//' @name postfix_normalized_similarity
 //' @title Normalized Postfix Similarity
 //' @description Calculates the normalized similarity between the postfixes of two strings.
 //'
@@ -73,8 +74,8 @@ double postfix_normalized_distance(std::string s1, std::string s2, double score_
 //'
 //' @examples
 //' postfix_normalized_similarity("string1", "string2")
+//' @export
 // [[Rcpp::export]]
 double postfix_normalized_similarity(std::string s1, std::string s2, double score_cutoff = 0.0) {
-  double score_hint = 0.0;
-  return rapidfuzz::detail::Postfix::normalized_similarity(s1, s2, score_cutoff, score_hint);
+  return rapidfuzz::postfix_normalized_similarity(s1, s2, score_cutoff);
 }